CP8117P12-231: Adding more pre-defined template in the category
JIRA link: https://pmc.cs.uwindsor.ca/jira/browse/CP8117P12-231 
Updated the xml file to include the add button and the list components to be dynamic based on data present in database. 
Updated the template.java file to get the data of categories from database and the add operation.